+++ packages/databyss-ui/components/Navigation/NavigationProvider/modalDict.js
@@ -1,24 +1,4 @@
-import {
-  View,
-  Button,
-  Text,
-  Grid,
-  ModalWindow,
-} from '@databyss-org/ui/primitives'
-
-const SourceModal = (visible, props, dismiss) => {
-  return (
-    <ModalWindow
-      visible={visible}
-      onDismiss={() => dismiss()}
-      title="Edit Source"
-      dismissChild="Save"
-      secondaryChild="Cancel"
-    >
-      <Text>{props}</Text>
-    </ModalWindow>
-  )
-}
+import SourceModal from '@databyss-org/ui/modules/SourcesValueList/SourceModal'
 
 export const modalDict = {
   SOURCE: SourceModal,
